fastapi-chatbot-app/backend/pyproject.toml
[project]
name = "fastapi-chatbot-backend"
version = "0.1.0"
description = "Backend service for the FastAPI chatbot application."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1,<0.105",
    "uvicorn>=0.24.0,<0.25",
    "httpx>=0.25.0,<0.26",
    "pydantic>=2.4.2,<3",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.0",
    "pandas>=2.1.3,<3",
    "pillow>=10.1.0,<11",
    "sentence-transformers>=2.2.2",
    "langchain>=0.3.3",
    "langchain-community>=0.3.2",
    "langchain-google-genai>=2.0.0",
    "qdrant-client>=1.7.0",
    "google-genai>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true
